Rename EmptyCriteria -> EmptyCondition in the namespace
