+++ src/Helper/SkipOnEmptyNormalizer.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Helper;
 
 use InvalidArgumentException;
-use Yiisoft\Validator\EmptyCriteria\NeverEmpty;
-use Yiisoft\Validator\EmptyCriteria\WhenEmpty;
+use Yiisoft\Validator\EmptyCondition\NeverEmpty;
+use Yiisoft\Validator\EmptyCondition\WhenEmpty;
 use Yiisoft\Validator\SkipOnEmptyInterface;
 
 use function is_callable;
