+++ tests/EmptyCriteria/WhenMissingTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\EmptyCriteria;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Validator\EmptyCriteria\WhenMissing;
+use Yiisoft\Validator\EmptyCondition\WhenMissing;
 use Yiisoft\Validator\Rule\Number;
 use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\Validator;
